+++ game_patch/sound/sound_ds.cpp
@@ -3,9 +3,12 @@
 #include <patch_common/ShortTypes.h>
 #include <patch_common/StaticBufferResizePatch.h>
 #include <patch_common/AsmOpcodes.h>
+#include <stb_vorbis.h>
 #include <algorithm>
 #include "../rf/sound/sound.h"
 #include "../rf/sound/sound_ds.h"
+#include "../rf/crt.h"
+#include "../rf/file/file.h"
 #include "../os/console.h"
 
 namespace rf
@@ -238,6 +241,110 @@ ConsoleCommand2 playing_sounds_cmd{
 };
 #endif
 
+struct MmioWrapper {
+    HMMIO hmmio = nullptr;
+    stb_vorbis* vorbis = nullptr;
+};
+
+FunHook<int(LPSTR, LONG, HMMIO*, WAVEFORMATEX**, WAVEFORMATEX**, LPMMCKINFO)> snd_mmio_open_hook{
+    0x00563370,
+    [](LPSTR filename, LONG offset, HMMIO* hmmio, WAVEFORMATEX **wfmt_orig, WAVEFORMATEX **wfmt_ds, LPMMCKINFO chunk_info) {
+        char *sound_name = filename - 0x100;
+        xlog::trace("Loading sound: %s", sound_name);
+        if (!strcmp(rf::file_get_ext(sound_name), ".ogg")) {
+            xlog::info("Loading Ogg Vorbis: %s", sound_name);
+            int file_size = rf::File{}.size(sound_name);
+            if (file_size <= 0) {
+                xlog::error("Cannot get file size: %s", filename);
+                return -1;
+            }
+            FILE *file = std::fopen(filename, "rb");
+            if (!file) {
+                xlog::error("Failed to open: %s", filename);
+                return -1;
+            }
+            std::fseek(file, offset, SEEK_SET);
+            int err = 0;
+            stb_vorbis* vorbis = stb_vorbis_open_file_section(file, 1, &err, NULL, file_size);
+            if (!vorbis) {
+                xlog::error("Failed to load Ogg Vorbis: %d", err);
+                return -1;
+            }
+
+            stb_vorbis_info info = stb_vorbis_get_info(vorbis);
+            xlog::info("Ogg Vorbis info: channels %d, sample rate %d", info.channels, info.sample_rate);
+
+            *wfmt_orig = reinterpret_cast<WAVEFORMATEX*>(rf::rf_malloc(sizeof(WAVEFORMATEX)));
+            *wfmt_ds = reinterpret_cast<WAVEFORMATEX*>(rf::rf_malloc(sizeof(WAVEFORMATEX)));
+            (*wfmt_orig)->wFormatTag = WAVE_FORMAT_PCM;
+            (*wfmt_orig)->nChannels = info.channels;
+            (*wfmt_orig)->wBitsPerSample = 16;
+            (*wfmt_orig)->nSamplesPerSec = info.sample_rate;
+            (*wfmt_orig)->nBlockAlign = (*wfmt_orig)->nChannels * (*wfmt_orig)->wBitsPerSample / 8;
+            (*wfmt_orig)->nAvgBytesPerSec = (*wfmt_orig)->nSamplesPerSec * (*wfmt_orig)->nBlockAlign;
+            (*wfmt_orig)->cbSize = 0;
+            **wfmt_ds = **wfmt_orig;
+            auto wrapper = new MmioWrapper;
+            wrapper->vorbis = vorbis;
+            *hmmio = reinterpret_cast<HMMIO>(wrapper);
+            return 0;
+        }
+        auto wrapper = new MmioWrapper;
+        *hmmio = reinterpret_cast<HMMIO>(wrapper);
+        return snd_mmio_open_hook.call_target(filename, offset, &wrapper->hmmio, wfmt_orig, wfmt_ds, chunk_info);
+    },
+};
+
+FunHook<int(HMMIO*, LPMMCKINFO, const MMCKINFO*)> snd_mmio_find_data_chunk_hook{
+    0x005635D0,
+    [](HMMIO* hmmio, LPMMCKINFO data_mmcki, const MMCKINFO* riff_mmcki) {
+        auto wrapper = reinterpret_cast<MmioWrapper*>(*hmmio);
+        if (wrapper->vorbis) {
+            xlog::info("Ogg Vorbis seek start");
+            stb_vorbis_seek_start(wrapper->vorbis);
+            stb_vorbis_info info = stb_vorbis_get_info(wrapper->vorbis);
+            float length = stb_vorbis_stream_length_in_seconds(wrapper->vorbis);
+            data_mmcki->cksize = info.sample_rate * info.channels * sizeof(short) * length;
+            return 0;
+        } else {
+            return snd_mmio_find_data_chunk_hook.call_target(&wrapper->hmmio, data_mmcki, riff_mmcki);
+        }
+    },
+};
+
+FunHook<int(HMMIO, unsigned, BYTE*, MMCKINFO*, int*)> snd_mmio_read_chunk_hook{
+    0x00563620,
+    [](HMMIO hmmio, unsigned buf_size, BYTE *buf, MMCKINFO *mmcki, int *bytes_read) {
+        auto wrapper = reinterpret_cast<MmioWrapper*>(hmmio);
+        if (wrapper->vorbis) {
+            stb_vorbis_info info = stb_vorbis_get_info(wrapper->vorbis);
+            int samples_read = stb_vorbis_get_samples_short_interleaved(wrapper->vorbis, info.channels, reinterpret_cast<short*>(buf), buf_size / 2);
+            *bytes_read = samples_read * info.channels * sizeof(short);
+            xlog::info("Ogg Vorbis: read %d samples", samples_read);
+            return 0;
+        }
+        return snd_mmio_read_chunk_hook.call_target(wrapper->hmmio, buf_size, buf, mmcki, bytes_read);
+    },
+};
+
+FunHook<void(HMMIO*)> snd_mmio_close_hook{
+    0x005636E0,
+    [](HMMIO *hmmio) {
+        auto wrapper = reinterpret_cast<MmioWrapper*>(*hmmio);
+        if (!wrapper) {
+            return;
+        }
+        if (wrapper->vorbis) {
+            xlog::info("Closing Ogg Vorbis stream");
+            stb_vorbis_close(wrapper->vorbis);
+        } else {
+            snd_mmio_close_hook.call_target(&wrapper->hmmio);
+        }
+        delete wrapper;
+        *hmmio = nullptr;
+    },
+};
+
 void snd_ds_apply_patch()
 {
     // Delete sounds with lowest volume when there is no free slot for a new sound
@@ -269,6 +376,12 @@ void snd_ds_apply_patch()
     snd_ds_play_3d_sig_generation_patch.install();
     snd_ds_play_music_sig_generation_patch.install();
     snd_ds_get_channel_hook.install();
+
+    // Add support for more formats
+    snd_mmio_open_hook.install();
+    snd_mmio_find_data_chunk_hook.install();
+    snd_mmio_read_chunk_hook.install();
+    snd_mmio_close_hook.install();
 
     // Commands
 #ifdef DEBUG